+++ pydirectus/translation.py
@@ -1,0 +1,7 @@
+from .session import Session
+
+
+class Translation:
+
+    def __init__(self, session: Session):
+        self.session = session