+++ pydirectus/typing.py
@@ -6,5 +6,5 @@ Alias: TypeAlias = str
 Other: TypeAlias = str
 Csv: TypeAlias = List[str]
 Hash: TypeAlias = str
-Uuid = uuid.UUID
+Uuid = str   # can't type as uuid.UUID because directus is str
 Json = Union[Dict[str, Any], List[Any]]